+++ eclair-core/src/test/scala/fr/acinq/eclair/db/FileBackupHandlerSpec.scala
@@ -19,12 +19,11 @@ package fr.acinq.eclair.db
 import java.io.File
 import java.sql.DriverManager
 import java.util.UUID
-
 import akka.testkit.TestProbe
 import fr.acinq.eclair.channel.ChannelPersisted
 import fr.acinq.eclair.db.Databases.FileBackup
 import fr.acinq.eclair.db.sqlite.SqliteChannelsDb
-import fr.acinq.eclair.wire.ChannelCodecsSpec
+import fr.acinq.eclair.wire.internal.channel.ChannelCodecsSpec
 import fr.acinq.eclair.{TestConstants, TestKitBaseClass, TestUtils, randomBytes32}
 import org.scalatest.funsuite.AnyFunSuiteLike
